Refine organization redirect handling
